+++ HISTORY.rst
@@ -13,6 +13,9 @@ Unreleased
 
 **Bugfixes**
 
+* ALCN-1196: Repeat each promotion fo Alcyon Group Id >= 100 in special promotions XML (ws02)
+* ALCN-1197: Change checksum calculation (use ids) for special promotions XML (wso2)
+
 **Build**
 
 **Documentation**